NotificationService was implemented
git-svn-id: svn+ssh://redwerk.com/home/svn/likelabs@5208 5beb62dc-ff0c-0410-899a-bd8837d9df5a

+++ server/sql/schema.sql
@@ -185,8 +185,8 @@ CREATE TABLE `notification_interval` (
      `id` BIGINT AUTO_INCREMENT NOT NULL,
      `event_type` TINYINT NOT NULL,
      `warning_type` TINYINT NOT NULL,
-     `email_interval` INT NOT NULL,
-     `sms_interval` INT NOT NULL,
+     `email_interval` TINYINT NOT NULL,
+     `sms_interval` TINYINT NOT NULL,
      CONSTRAINT `PK_notification_interval` PRIMARY KEY (`id`)
 ) ENGINE=InnoDB, DEFAULT CHARSET=utf8;
